+++ CHANGELOG.md
@@ -7,10 +7,15 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 
 
-## [1.4.2 / 5.59.2] - 2022-10-??
+## [1.4.2 / 5.59.2] - 2022-10-10
 
 ### Added
 - added tooltips to resource access modes [#2300](https://github.com/sandboxie-plus/Sandboxie/issues/2300)
+- added UI option to control ApplyElevateCreateProcessFix [#2302](https://github.com/sandboxie-plus/Sandboxie/issues/2302)
+- added message 2226 issued when a process may need 'ApplyElevateCreateProcessFix=y'
+
+### Changed
+- Moved Restrictions tab from the new Security page back to the General page
 
 ### Fixed
 - fixed wrong button captions in the breakout process page
@@ -21,6 +26,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - fixed miscellaneous minor issues [#2301](https://github.com/sandboxie-plus/Sandboxie/issues/2301)
 - fixed issue with MSEdge introduced in 106.x [#2325](https://github.com/sandboxie-plus/Sandboxie/issues/2325)
 - fixed vivaldi hooking issue, UseVivaldiWorkaround is no longer needed [#1783](https://github.com/sandboxie-plus/Sandboxie/issues/1783)
+- fixed issues with misc advanced options page [#2315](https://github.com/sandboxie-plus/Sandboxie/issues/2315)